+++ client/src/store.js
@@ -1,0 +1,9 @@
+import { configureStore } from "@reduxjs/toolkit";
+import rootReducer from "./reducers";
+
+const store = configureStore({
+  reducer: rootReducer,
+  // Additional middleware and configuration if needed
+});
+
+export default store;
